Broke out ApplicationList from RecruiterView
+ Added ApplicationList.tsx, which contains most of the contents from RecruiterView.tsx.
Reason: modularity, also makes RecruiterView simpler.
