+++ client/src/view/styles/ApplicationList.css
